Fix for rounding issue when calculating step angle.
This caused a rendering gap where in some cases the last fin to be drawn would not touch the first fin (this is visible in the demo screenshot).

+++ HZActivityIndicator/HZActivityIndicatorView.m
@@ -185,7 +185,7 @@
 
 - (void)setSteps:(NSUInteger)steps
 {
-    _anglePerStep = (360/steps) * M_PI / 180;
+    _anglePerStep = (360. / (CGFloat)steps) * M_PI / 180.;
     _steps = steps;
     [self setNeedsDisplay];
 }
